Separated game logic and web API
