+++ internal/game/lobby.go
@@ -1,8 +1,7 @@
-package server
+package game
 
 import (
 	"sync"
-	"time"
 
 	"github.com/google/uuid"
 )
@@ -23,20 +22,6 @@ func (m Mark) String() string {
 	default:
 		return "?"
 	}
-}
-
-type PlayerScores struct {
-	WinCount  int `json:"winCount"`
-	LoseCount int `json:"loseCount"`
-	TieCount  int `json:"tieCount"`
-}
-
-type Player struct {
-	Name           string        `json:"name"`
-	Token          string        `json:"token"`
-	BanTimestamp   *time.Time    `json:"banTimestamp"`
-	DateOfRegister time.Time     `json:"dateOfRegister"`
-	Scores         *PlayerScores `json:"scores"`
 }
 
 type Lobby struct {
