+++ internal/game/lobby_test.go
@@ -1,4 +1,4 @@
-package server
+package game
 
 import (
 	"encoding/json"
