+++ internal/game/tictactoe.go
@@ -1,4 +1,4 @@
-package server
+package game
 
 import "errors"
