+++ internal/server/adminApi.go
@@ -4,15 +4,6 @@ import (
 	"net/http"
 )
 
-func (ttts *TicTacToeServer) GenerateGrid() TicTacToeGrid {
-
-	grid := TicTacToeGrid{
-		XMarks: []TicTacToeMark{},
-		OMarks: []TicTacToeMark{},
-	}
-	return grid
-}
-
 func (ttts *TicTacToeServer) HandleAdminListPlayers(w http.ResponseWriter, r *http.Request) {
 	r.ParseForm()
 	adminToken := r.Form.Get("admintoken")
