+++ internal/server/webinterface.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/gorilla/securecookie"
 	"github.com/gorilla/sessions"
+	"github.com/the-developer-guy/tic-tac-multiplayer/internal/game"
 )
 
 var (
@@ -21,8 +22,8 @@ var (
 )
 
 type FetchScores struct {
-	Name  string       `json:"name"`
-	Score PlayerScores `json:"scores"`
+	Name  string            `json:"name"`
+	Score game.PlayerScores `json:"scores"`
 }
 
 type LoginStruct struct {
